refactor: replace Config with Workflow in API and loader implementations
