+++ crates/forge_domain/src/lib.rs
@@ -4,7 +4,7 @@ mod attachment;
 mod chat_request;
 mod chat_response;
 mod compaction;
-mod config;
+mod workflow;
 mod context;
 mod conversation;
 mod env;
@@ -32,7 +32,6 @@ mod tool_definition;
 mod tool_name;
 mod tool_result;
 mod tool_usage;
-mod workflow;
 
 pub use agent::*;
 pub use api::*;
@@ -40,7 +39,7 @@ pub use attachment::*;
 pub use chat_request::*;
 pub use chat_response::*;
 pub use compaction::*;
-pub use config::*;
+pub use workflow::*;
 pub use context::*;
 pub use conversation::*;
 pub use env::*;
@@ -67,4 +66,3 @@ pub use tool_definition::*;
 pub use tool_name::*;
 pub use tool_result::*;
 pub use tool_usage::*;
-pub use workflow::*;
